requirements.txt
spotipy
fastapi
python-dotenv
Flask
Flask-Session